update: improve and make ui consistent
