+++ src/components/Evaluation.js
@@ -5,30 +5,20 @@ import Card, { CardHeader, CardActions, CardContent } from 'material-ui/Card'
 import Button from 'material-ui/Button'
 import Typography from 'material-ui/Typography'
 
-import { withStyles } from 'material-ui/styles'
-
-const styles = theme => ({
-	card: {
-		margin: theme.spacing.unit * 2,
-        width: 400
-	},
-    card_content: {
-        paddingTop: 0,
-        paddingBottom: 0
-    },
-    card_actions: {
-        padding: [[8, 0]]
-    }
-})
-
 const Evaluation = ({ classes, data, handleDelete }) => {
         
     return (
         <Card className={classes.card}> 
-            <CardHeader title={ data.metric_name } />
+            <CardHeader title={
+                <Typography variant='title'> { data.metric_name }</Typography>
+            }/>
             <CardContent className={ classes.card_content }>
                 <Typography component='p'>
-                    The model trained with job id { data.job_id } achieved a performance of { data.metric_value } when evaluated on this metric
+                    The model trained with job id &nbsp;
+                    <code>{ data.job_id }</code> &nbsp;
+                    achieved a performance of &nbsp;
+                    <code>{ data.metric_value }</code> &nbsp;
+                    when evaluated on this metric
                 </Typography>
             </CardContent>
             <CardActions className={ classes.card_actions }>
@@ -46,4 +36,4 @@ const Evaluation = ({ classes, data, handleDelete }) => {
     )
 }
 
-export default withStyles(styles)(Evaluation)
+export default Evaluation
